requirements.txt
fastapi==0.78.0
psycopg2==2.9.3
SQLAlchemy==1.4.39
alembic==1.8.0
sqlalchemy-utils==0.38.2
python-dotenv==0.20.0